chore: move data classes to correct namespaces
