--- Core/NosSmooth.Game/Data/Characters/Family.cs
@@ -1,9 +1,0 @@
-﻿// 
-// Family.cs
-// 
-// Copyright (c) Christofel authors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace NosSmooth.Game.Data.Character;
-
-public record Family();
